Add authorization

+++ app/Http/Controllers/DocumentController.php
@@ -7,6 +7,10 @@ use Illuminate\Http\Request;
 
 class DocumentController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth')->except(['index', 'show']);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -40,8 +44,8 @@ class DocumentController extends Controller
             'title' => ['required', 'min:6']
         ]);
 
-        $attributes['user_id'] = 1;
-        
+        $attributes['user_id'] = auth()->id();
+
         Document::create($attributes);
 
         return redirect('/documents');
@@ -66,6 +70,8 @@ class DocumentController extends Controller
      */
     public function edit(Document $document)
     {
+        $this->authorize('update', $document);
+
         return view('documents.edit', compact('document'));
     }
 
@@ -78,6 +84,8 @@ class DocumentController extends Controller
      */
     public function update(Request $request, Document $document)
     {
+        $this->authorize('update', $document);
+
         $document->update(request()->validate([
             'title' => ['required', 'min:6']
         ]));
@@ -93,6 +101,8 @@ class DocumentController extends Controller
      */
     public function destroy(Document $document)
     {
+        $this->authorize('update', $document);
+        
         $document->delete();
 
         return redirect('/documents');
